requirements.txt
acryl-datahub==0.10.1.1
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.3
async-timeout==4.0.2
attrs==22.2.0
avro==1.10.2
avro-gen3==0.7.10
build==0.10.0
cached-property==1.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-default-group==1.2.2
click-spinner==0.1.10
cloudpickle==2.2.1
contourpy==1.0.7
-e git+https://github.com/hariishaa/datahub-mlflow-source.git@f8935fdd653275945bdac1696bfa0083e7ba8455#egg=custom_source_examples
cycler==0.11.0
databricks-cli==0.17.6
Deprecated==1.2.13
docker==6.0.1
entrypoints==0.4
expandvars==0.9.0
Flask==2.2.3
fonttools==4.39.3
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
gunicorn==20.1.0
humanfriendly==10.0
idna==3.4
ijson==3.2.0.post0
importlib-metadata==6.3.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonref==1.1.0
jsonschema==4.17.3
kiwisolver==1.4.4
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mixpanel==4.10.0
mlflow==2.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
progressbar2==4.2.0
protobuf==4.22.3
psutil==5.9.4
pyarrow==11.0.0
pydantic==1.10.7
PyJWT==2.6.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-utils==3.5.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
querystring-parser==1.2.4
ratelimiter==1.2.0.post0
requests==2.28.2
requests-file==1.5.1
scikit-learn==1.2.2
scipy==1.10.1
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
SQLAlchemy==2.0.9
sqlparse==0.4.3
tabulate==0.9.0
termcolor==2.2.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
websocket-client==1.5.1
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.8.2
zipp==3.15.0
